# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "C2D Admin API",
    "version": "1.0.0",
    "description": "API для админки ИИ-агента и рассылки сообщений"
  },
  "servers": [
    {"url": "/api", "description": "API Server"}
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Авторизация пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "admin@c2d.ru"},
                  "password": {"type": "string", "example": "Admin2026!"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Успешная авторизация"},
          "422": {"description": "Неверные учетные данные"}
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Выход из системы",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Успешный выход"}}
      }
    },
    "/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Получить текущего пользователя",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Данные пользователя"}}
      }
    },
    "/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Список рекламных кампаний",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "status", "in": "query", "schema": {"type": "string"}, "description": "Фильтр по статусу"}
        ],
        "responses": {"200": {"description": "Список кампаний"}}
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Создать рекламную кампанию",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string", "example": "Новогодняя акция"},
                  "description": {"type": "string"},
                  "product_id": {"type": "integer"},
                  "trigger_action": {"type": "string"},
                  "target_url": {"type": "string"},
                  "max_retries": {"type": "integer", "example": 3}
                }
              }
            }
          }
        },
        "responses": {"201": {"description": "Кампания создана"}}
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Получить кампанию по ID",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Данные кампании"}}
      },
      "put": {
        "tags": ["Campaigns"],
        "summary": "Обновить кампанию",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Кампания обновлена"}}
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Удалить кампанию",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Кампания удалена"}}
      }
    },
    "/campaigns/{id}/launch": {
      "post": {
        "tags": ["Campaigns"],
        "summary": "Запустить кампанию",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Кампания запущена"}}
      }
    },
    "/campaigns/{id}/pause": {
      "post": {
        "tags": ["Campaigns"],
        "summary": "Приостановить кампанию",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Кампания приостановлена"}}
      }
    },
...
  }
}